+++ Sofa/Component/Topology/Container/Dynamic/CMakeLists.txt
@@ -42,12 +42,12 @@ set(HEADER_FILES
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TetrahedronSetTopologyAlgorithms.h
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TetrahedronSetTopologyContainer.h
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TetrahedronSetTopologyModifier.h
-    ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleModifiers.h
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSetGeometryAlgorithms.h
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSetGeometryAlgorithms.inl
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSetTopologyAlgorithms.h
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSetTopologyContainer.h
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSetTopologyModifier.h
+    ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSubdividers.h
 )
 
 set(SOURCE_FILES
@@ -68,11 +68,11 @@ set(SOURCE_FILES
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/QuadSetTopologyModifier.cpp
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TetrahedronSetGeometryAlgorithms.cpp
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TetrahedronSetTopologyContainer.cpp
-    ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TetrahedronSetTopologyModifier.cpp
-    ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleModifiers.cpp
+    ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TetrahedronSetTopologyModifier.cpp    
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSetGeometryAlgorithms.cpp
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSetTopologyContainer.cpp
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSetTopologyModifier.cpp
+    ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/TriangleSubdividers.cpp
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/DynamicSparseGridGeometryAlgorithms.cpp
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/DynamicSparseGridTopologyContainer.cpp
     ${SOFACOMPONENTTOPOLOGYCONTAINERDYNAMIC_SOURCE_DIR}/DynamicSparseGridTopologyModifier.cpp
